Check OLD_SOUND_FLAG or SOUND_FLAG if V3 #40
Check for sound now checks for either `OLD_SOUND_FLAG` or `SOUND_FLAG` if the story version is V3.

Inform 6 sets SOUND_FLAG when compiling to V3. There will soon be an option to set and clear whatever flags you please, but the default will still be SOUND_FLAG. This covers all bases, in much the same way the current build of Frotz handles this.

#40